backend/requirements.txt
Django==2.2.16
django-cors-headers==3.11.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
pytz==2021.3
drf-extra-fields==3.4.0
Pillow
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
oauthlib==3.2.0
psycopg2-binary==2.8.6
python-dotenv